requirements.txt
SQLAlchemy~=2.0.41
dotenv~=0.9.9
python-dotenv~=1.1.1
aio-pika~=9.5.5
Telethon~=1.37.0
PySocks~=1.7.1
asyncpg~=0.30.0
